requirements.txt
aiofiles==24.1.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.1
asgiref==3.7.2
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
azure-monitor-opentelemetry-exporter==1.0.0b32
azure-servicebus==7.11.4
azure-storage-blob==12.24.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.4
filelock==3.16.1
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.0
Flask-Cors==4.0.2
flask-restx==1.3.0
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
halo==0.0.31
httpcore==1.0.6
httpx==0.27.2
idna==3.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
log-symbols==0.0.14
MarkupSafe==2.1.3
marshmallow==3.23.1
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
numpy==2.1.3
oauthlib==3.2.2
openai==1.54.3
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
packaging==24.2
pandas==2.2.3
pillow==10.4.0
portalocker==2.8.2
promptflow==1.16.1
promptflow-core==1.16.1
promptflow-devkit==1.16.1
promptflow-tools==1.4.0
promptflow-tracing==1.16.1
prompty==0.1.37
protobuf==5.28.3
psutil==5.9.8
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydash==7.0.7
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.21.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
spinners==0.0.24
SQLAlchemy==2.0.36
starlette==0.41.2
strictyaml==1.7.3
tabulate==0.9.0
termcolor==2.5.0
tiktoken==0.8.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.1.0
waitress==2.1.2
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.21.0
